+++ web/src/app/gameplay/page.tsx
@@ -1,3 +1,5 @@
+'use client';
+
 import { useState, useEffect, useCallback } from 'react'
 import Link from 'next/link'
 import { useCurrentAccount, useSignAndExecuteTransaction } from '@mysten/dapp-kit'
