+++ web/src/app/providers.tsx
@@ -1,5 +1,6 @@
-import '@/styles/globals.css'
-import type { AppProps } from 'next/app'
+'use client';
+
+import { ReactNode, useState } from 'react'
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query'
 import {
   createNetworkConfig,
@@ -7,7 +8,6 @@ import {
   WalletProvider,
 } from '@mysten/dapp-kit'
 import { getFullnodeUrl } from '@mysten/sui/client'
-import '@mysten/dapp-kit/dist/index.css'
 
 const { networkConfig } = createNetworkConfig({
   localnet: { url: getFullnodeUrl('localnet') },
@@ -16,15 +16,13 @@ const { networkConfig } = createNetworkConfig({
   mainnet: { url: getFullnodeUrl('mainnet') },
 })
 
-const queryClient = new QueryClient()
+export function Providers({ children }: { children: ReactNode }) {
+  const [queryClient] = useState(() => new QueryClient())
 
-export default function App({ Component, pageProps }: AppProps) {
   return (
     <QueryClientProvider client={queryClient}>
       <SuiClientProvider networks={networkConfig} defaultNetwork="testnet">
-        <WalletProvider autoConnect>
-          <Component {...pageProps} />
-        </WalletProvider>
+        <WalletProvider autoConnect>{children}</WalletProvider>
       </SuiClientProvider>
     </QueryClientProvider>
   )
