+++ web/src/components/SwimmingPool.tsx
